tg-bot/requirements.txt
requests>=2.32.3
python-dotenv>=1.1.0
aiogram>=3.20.0
